fix: architecture
