--- component/theme/theme.tsx
@@ -1,11 +1,0 @@
-import {createTheme} from "@mui/material";
-
-const CodePieTheme = createTheme({
-    palette: {
-        primary: {
-            main: '#EC8000',
-        }
-    }
-});
-
-export default CodePieTheme;
