+++ pages/feeds/index.tsx
@@ -1,0 +1,5 @@
+import FeedsLayout from '../../components/feeds/FeedsLayout';
+
+export default function FeedsPage() {
+  return <FeedsLayout />;
+}
