--- stories/pages/feed_list/side_card.stories.tsx
@@ -1,12 +1,0 @@
-import {ComponentMeta, ComponentStory} from "@storybook/react";
-import SideCard from "../../../pages/feed_list/_side_card";
-
-export default {
-    title: 'codepie/SideCard',
-    component: SideCard,
-} as ComponentMeta<typeof SideCard>;
-
-const Template: ComponentStory<typeof SideCard> = () => <SideCard/>;
-
-export const Basic = Template.bind({});
-Basic.args = {};
