+++ styles/theme.ts
@@ -1,0 +1,11 @@
+import { createTheme } from '@mui/material';
+
+const CodePieTheme = createTheme({
+  palette: {
+    primary: {
+      main: '#EC8000',
+    },
+  },
+});
+
+export default CodePieTheme;
